requirements.txt
Django==3.1.1
Pillow==8.2.0
